Make Mining Challenge component more modular (#535)
* Commmit-1: Introduce card and Block counter component

- Use it to create the first component of mining-1 page

* Commit 2: Updated progress.tsx to have two variants

- Replaced first component with Card

* Commit-3: Introduced HashDisplayer and NonceCounter component

- Replaced 2nd component with Card

* Commit-4: Added Highlight feature to Card

* Commit-5: Modularised MiningStatisticHash component

- Introduced, TitleCard, HashFrequency, and StartButton component.

* Commit-6: Replace MiningStatistic with modular implementation

* Commit-7: Cleanup. Remove redundant files

* Commit-8: Address comments

* Commit-9: Simplify Progress logic

* Commit-10: Fix issue related to size of dual Cards
