+++ ui/common/index.ts
@@ -8,7 +8,11 @@ export { default as Tip } from './Tip'
 export { default as Input } from './Input'
 export { default as Hasher } from './Hasher'
 export { default as TextImage } from './TextImageLayout'
-export { default as MiningStatistic } from './MiningStatistic'
 export { default as ProgressBar } from './ProgressBar'
-export { default as MiningStatisticNonce } from './MiningStatisticsNonce'
-export { default as MiningStatisticHash } from './MiningStatisticsHash'
+export { default as Card } from './Card'
+export { default as BlockCounter } from './BlockCounter'
+export { default as NonceCounter } from './NonceCounter'
+export { default as HashDisplayer } from './HashDisplayer'
+export { default as TitleCard } from './TitleCard'
+export { default as HashFrequency } from './HashFrequency'
+export { default as StartButton } from './StartButton'
